pyproject.toml
[project]
name = "ai-engineer-capstone"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.13.2",
    "fastapi>=0.128.0",
    "mcp[cli]>=1.26.0",
    "numpy>=2.4.1",
    "openai>=2.15.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "ruff>=0.14.14",
    "sqlalchemy>=2.0.0",
    "structlog>=25.5.0",
    "tiktoken>=0.12.0",
    "uvicorn[standard]>=0.40.0",
]
